requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.3
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pyodbc==5.1.0
streamlit==1.32.2
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
